Started changing structure
